Support CD/DVD changing during package installation.
For this to be possible, the installation CD/DVD has to be unmounted in
the D-I environment, which means that during most of pkgsel (including
pre-pkgsel.d scripts) the installation of udebs is not possible. The
installation CD/DVD is reloaded before the end of pkgsel to ensure
following components can again install udebs.



git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@50017 48c42b26-1dd6-0310-b98f-a58d8bce7237