Clean up obsolete partition directories
When changes to a device are committed using parted, we currently do not
remove any obsolete partition directories in /var/lib/partman/devices/$dev.
This stale data is confusing and could potentially lead to problems.

Introduce a new function device_cleanup_partitions() in the commit.sh
function library to ensure the partition info is consistent.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@50654 48c42b26-1dd6-0310-b98f-a58d8bce7237