Factor out new function refine_locale
This mostly just makes the code a lot better readable.

The old code contained an otherwise unused variable LOG, which means
that the test that seems to have been intended to avoid "refining" the
locale if it was preseeded, always had the same outcome.
The dead code has been removed and the comments have been updated to
indicate that this should possibly be revisited at some point.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@51840 48c42b26-1dd6-0310-b98f-a58d8bce7237