Add ath5k wireless module
WEP support requires the ecb and arc4 modules. As these are only
used for wireless, they can be added in nic-wireless-modules.

Since ecb depends on blkcipher that module now needs to be split
out from crypto-modules into a new udeb crypto-core-modules on which
both nic-wireless and crypto depend. At the moment only architectures
that have nic-wireless need this new udeb.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@52373 48c42b26-1dd6-0310-b98f-a58d8bce7237