Put strings relevant to the newt entropy plugin in new templates
debconf/entropy/text/action and debconf/entropy/text/help, shipped in
the cdebconf-newt-entropy package are now used by the plugin to get the
strings that will be prompted by the user.

The short description of the entropy question will now be prepended to the
help text, thus removing the need to be specific on how to produce more
entropy.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@52550 48c42b26-1dd6-0310-b98f-a58d8bce7237