Avoid unnecessary untranslated values because of PO file conflicts
Currently conficts can arise, for example in the translation of
continents, when the localechooser and iso-codes PO files are merged,
leading to untranslated values in the dialogs.

This is solved in two ways:
- remove the gettext call for Antarctica in mktemplates.continents as
  that is already included in the ISO-3166 file
- use the --use-first option when msgcat'ing the PO files (to avoid the
  same problem in the future)

The iso-codes file is listed first so that the more official translation
is preferred when there are differences.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@52639 48c42b26-1dd6-0310-b98f-a58d8bce7237