Also call the cleanup function for "Go back"
The cleanup function of the GTK+ plugin were previously only called when
the "Continue" button was selected.  This resulted in the FIFO being kept
when "Go back" was selected.

The plugin setter function is *not* called when "Go back" is selected.
So we now use the destruction of the GTK+ widgets as the source for
calling back the cleanup function.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@53320 48c42b26-1dd6-0310-b98f-a58d8bce7237