Fix partition numbering when confirming changes
In confirm_changes(), the control flow for displaying the list of
partitions was changed when support for multipath devices was added.
Before that $partdesc was redefined for every partitions, either for
device mapper devices or normal ones.

The multipath changes introduced a conditionnal branch testing for the
emptyness of $partdesc.  As it was previously set for the first partition,
the second time, the template variable was not previously substituted.

We now explicitely reset $partdesc for each partition in order to get the
proper behaviour.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@53469 48c42b26-1dd6-0310-b98f-a58d8bce7237