Try to install the new ixp4xx-microcode package
Fall back to copying microcode files only if the installation fails.

The package will fail to install most of the time now, since non-free is
not available by default. This at least gets a cleaner microcode
installation if the user does choose to use non-free. 

(Also, In the future, hw-detect will notice missing firmware requests, and
may enable using non-free, so this future-proofs the microcode installation.)

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@53811 48c42b26-1dd6-0310-b98f-a58d8bce7237