move upgrade to before pre-pkgsel
Specifically, I want it to run before the popcon question, because this
way there's not a long wait for the upgrade in between the popcon and tasksel
questions.

This also avoids needing to upgrade a few things like popcon and
laptop-detect, which can't hurt.

Moving the upgrade into a pre-pkgsel script itself would be conceptually
cleaner, but such scripts have no way to fail the whole pkgsel run, which
needs to happen if an upgrade fails. Also, the code sets variables that are
used later in pkgsel.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@54088 48c42b26-1dd6-0310-b98f-a58d8bce7237