Harmonize size specification used for new LVs
Previously, the size given when creating a new Logical Volume was given
directly to lvcreate.  The later uses size units as multiple of 1024
whereas the rest of partman uses size units as multiple of 1000.

In order to solve this, we now gives LVM extents to lvcreate instead of
size.  The following was modified in lib/lvm-base.sh:
 - lv_create() now uses extents instead of size,
 - lvm_size_from_human() has been renamed to
   lvm_extents_from_human(),
 - As extent sizes are specific to a given Volume
   Group, lvm_extents_from_human() take the VG name as its first
   argument.

do_lv_create() in partman-lvm and perform_recipe_by_lvm in
partman-auto-lvm were modified accordingly.

(Closes: #411943)


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@54527 48c42b26-1dd6-0310-b98f-a58d8bce7237