Remove MD devices when creating a new disk label
In create_new_label(), we now remove MD devices like it was already done
for LVM.  As partman-md now scans for existing MD devices during
partman initialization, it is required in order to make automatic
partitioning working on disks which was previously used for an MD device.

A small library, lib/md-remove.sh has been added to partman-md in order to
do so.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@54714 48c42b26-1dd6-0310-b98f-a58d8bce7237