Restart partman once for all devices in prepare_new_labels
device_remove_lvm() and device_remove_md() now returns 99 when partman
needs to be restarted.  This allows prepare_new_labels() to restart
partman only once for all devices.

This change is necessary to allow proper handling of automatic
partitioning of multiple devices having RAID or LVM partitions.

Prior to this change, we had the following scenario:
 1. We create a new label is created on the first disk, without
    commiting the change.
 2. In order to create a new label on the second disk, we remove an LVM
    partition.
 3. As we have removed an LVM partition, we restart partman.
 4. While initializing, partman gets rid of its previous data and rescan
    the partitions.
 5. The state of the first disk is thus restored, and automatic
    partitioning is most likely to fail.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@55259 48c42b26-1dd6-0310-b98f-a58d8bce7237