Dynamically adjust priority at which menu is displayed
Instead of lowering the debconf priority after the user backs up from
a question, match the priority at which the menu is displayed to the
current debconf priority.

This means that change debconf priority will now in most cases display
the actual current debconf priority and in general behavior becomes
more consistent.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@56986 48c42b26-1dd6-0310-b98f-a58d8bce7237