Ensure ehci_hcd is loaded before [ou]hci_hcd for the target system
When these modules are loaded in the wrong order, the USB busses
will be reset which can lead to harmless but ugly and unnecessary
kernel errors, for example:
usb 3-1: device not accepting address 2, error -71
hub 3-0:1.0: unable to enumerate USB device on port 1

The kernel itself cannot ensure the modules are loaded in the correct
order, so to avoid these errors we use modprobe to do this. But only
if the installer has loaded both ehci_hcd and either [ou]hci_hcd.

This also avoids a warning printed by 2.6.27 and higher kernels:
Warning! ehci_hcd should always be loaded before uhci_hcd and ohci_hcd, not after

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@57868 48c42b26-1dd6-0310-b98f-a58d8bce7237