svn-all-fast-export patch for proper branch handling
This is a really quick and dirty and ugly fix for a bug that broke a lot
of d-i tags. If a tag branches from an older svn rev than the current one,
we need to somehow tell git-fast-import what rev it branched from. In order
to do that, we need to mark every revision we can, and record revisions we
were unable to mark in a lookup table so we can use the closest (previous)
revision if necessary.


git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@60075 48c42b26-1dd6-0310-b98f-a58d8bce7237