Run depmod -a before configuring dasd devices
Not strictly needed for the udeb itself, but makes it possible for users
to mount a dasd partition immediately afterwards as it makes file system
modules available. And it is needed for rescue mode.
The same is also done in hw-detect, so this improves consistency.

Also renames dasd to dasd-config as it is now installed as a separate
program and just dasd is not very clear.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@60208 48c42b26-1dd6-0310-b98f-a58d8bce7237