Implement option to not install recommended packages by default
The default is to install Recommends, but this offers experienced
users the option of installing a system with APT configured to not
install recommended packages.

To activate the option the user can either preseed the template
'base-installer/install-recommends' to false or set the option at
the boot prompt using the alias 'recommends=false'.

Note that using this option can result in an incomplete system and
additional packages may need to be installed manually to obtain
expected functionality.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@60485 48c42b26-1dd6-0310-b98f-a58d8bce7237