Change default video mode for i386/amd64 to vga=788 for the "newt" frontend
This forces the framebuffer to VESA and will thereby solve problems reported
on various HP notebooks with the vga16fb driver (#471505).

It also increases the screen size (to 600x800, which is still quite
conservative) and thus allows more information to be displayed. Note that
dialogs should continue to be designed to work with 480x640 as other
architectures may still use that as a default, and for network-console.

vga=788 is the same video mode we use for the gtk frontend and as we've seen
very few reports of that failing the change should be safe, but it is still
somewhat experimental.



git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@60979 48c42b26-1dd6-0310-b98f-a58d8bce7237