Allow separate preseeding of language and country
Currently users can only preseed the combination of language, country
and locale by preseeding debian-installer/locale. The problem with
this is that with this mechanism some valid combinations cannot be
preseeded.

Allow users to also separately preseed the language and country templates.
If *only* locale is preseeded the old functionality is preserved.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@61309 48c42b26-1dd6-0310-b98f-a58d8bce7237