Set d-i debconf variables needed to configure console-setup in target environment
Make sure debian-installer debconf variables needed by console-setup
are available in the target environment before installing the package.

This may not be the best possible solution, but it works and avoids the
problem that acpi-support-base pulls in console-{common,data} if the
installation of console-setup is postponed.

The template debian-installer/charmap is not set as its status in unclear
after the upload of console-setup 1.47.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@61330 48c42b26-1dd6-0310-b98f-a58d8bce7237