check-missing-firmware: handle case where driver does not register itself
Not all drivers register themselves if firmware is missing which means
the current code does not find the module name and the user isn't asked
if he wants to load the firmware. In such cases fall back to determining
the module using the device's modalias.

Thanks to Jérémy Bobbio for providing the patch; minor changes by FJP.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@61891 48c42b26-1dd6-0310-b98f-a58d8bce7237