Extend selection of preferred locale
Also allow selection of a "preferred locale", but only at medium priority,
when the user selects a combination of language and country for which a
locale is defined if multiple UTF-8 locales exist for the language.

Example:
A Brit living in Australia who prefers using en_GB.UTF-8 over en_AU.UTF-8.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@61951 48c42b26-1dd6-0310-b98f-a58d8bce7237