Avoid display of untranslated country codes
Improve generation of country selection templates to avoid display of
country codes in case translations are missing for some countries for a
particular language, or if no iso-codes translation exists for a language.

An extra step of template trickery is added which creates a template file
with language codes in Choices-C and English translations in Choices.
Only after that other translations are merged. In one respect it's even
simpler than what we had as we no longer need to create countrycode based
PO files for all translations.

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@62078 48c42b26-1dd6-0310-b98f-a58d8bce7237