Add a reversed Package/Provides mapping.
edos wasn't previously mentioning virtual packages. Since dose does that
now, build a mapping to perform provided/provider lookups. This lowers
the likelihood of having separate graphs when that shouldn't be the case.

An example can be seen with ia64 in wheezy (e.g. with the installed-based
and mounted-partitions virtual packages).

git-svn-id: svn+ssh://svn.debian.org/svn/d-i/trunk@70003 48c42b26-1dd6-0310-b98f-a58d8bce7237